src/the_grand_house_of_avneesh/summary/requirements.txt
# Core dependencies
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.3
langchain-aws==0.1.0
boto3==1.34.0
python-dotenv==1.0.0

# For development/testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
